add nearest neighbour benchmark

+++ include/path_finder/storage/DataConfig.h
@@ -22,8 +22,8 @@ struct GridMapEntry {
   std::pair<NodeId, NodeId> pointerPair;
 };
 struct GridMapEntries {
-  int latStretchFactor;
-  int lngStretchFactor;
+  double latStretchFactor;
+  double lngStretchFactor;
   std::vector<GridMapEntry> gridMapEntries;
 };
